Refactors CustomListEntriesEditor into three new functional components
