+++ src/components/CustomListEditorHeader.tsx
@@ -1,27 +1,24 @@
 import * as React from "react";
 import TextWithEditMode from "./TextWithEditMode";
-import { CollectionData } from "opds-web-client/lib/interfaces";
 import { Button } from "library-simplified-reusable-components";
 import { Entry } from "./CustomListEntriesEditor";
 
 export interface CustomListEditorHeaderProps {
   draftTitle: string;
-  list?: CollectionData;
   listId?: string | number;
   hasListInfoChanged: boolean;
   listEntries: Entry[];
+  cancelClicked: () => void;
   setDraftTitle: (title) => void;
-  setDraftEntries: (entries) => void;
   saveFormData: () => void;
 }
 
 export default function CustomListEditorHeader({
   draftTitle,
-  list,
   listId,
   hasListInfoChanged,
   setDraftTitle,
-  setDraftEntries,
+  cancelClicked,
   saveFormData,
   listEntries,
 }: CustomListEditorHeaderProps) {
@@ -38,11 +35,6 @@ export default function CustomListEditorHeader({
 
   const setNewTitleOnState = (newTitle): void => {
     setDraftTitle(newTitle);
-  };
-
-  const resetState = (): void => {
-    setDraftTitle(list ? list.title : "");
-    setDraftEntries(list ? list.books : []);
   };
 
   return (
@@ -68,7 +60,7 @@ export default function CustomListEditorHeader({
         />
         <Button
           className="inverted"
-          callback={resetState}
+          callback={cancelClicked}
           disabled={!hasListInfoChanged}
           content="Cancel Changes"
         />
